double -> long double

+++ Point.cpp
@@ -12,10 +12,10 @@ Point operator-(Point one, Point other) {
     return one + (-other);
 }
 
-Point operator*(double coeff, Point point) {
+Point operator*(long double coeff, Point point) {
     return Point(coeff * point.x, coeff * point.p);
 }
 
-Point operator/(Point point, double coeff) {
+Point operator/(Point point, long double coeff) {
     return Point(point.x / coeff, point.p / coeff);
 }
