+++ Point.hpp
@@ -6,12 +6,12 @@
 #include <cmath>
 
 struct Point {
-    double x;
-    double p;
+    long double x;
+    long double p;
 
-    Point(double x_ = 0, double p_ = 0) : x(x_), p(p_) {};
+    Point(long double x_ = 0, long double p_ = 0) : x(x_), p(p_) {};
 
-    double linf_norm() {
+    long double linf_norm() {
         return std::max(std::fabs(x), std::fabs(p));
     }
 
@@ -19,8 +19,8 @@ struct Point {
     friend Point operator-(Point one);
     friend Point operator-(Point one, Point other);
 
-    friend Point operator*(double coeff, Point point);
-    friend Point operator/(Point point, double coeff);
+    friend Point operator*(long double coeff, Point point);
+    friend Point operator/(Point point, long double coeff);
 };
 
 #endif
